initials 2]

+++ app/Http/Controllers/SupplierController.php
@@ -25,6 +25,7 @@ class SupplierController extends Controller
      */
     public function create()
     {
+        $this->authorize('user', 'admin', SupplierPolicy::class);
         return view('supplier.addsupplier');
     }
 
@@ -36,6 +37,7 @@ class SupplierController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('user', 'admin', SupplierPolicy::class);
         $supplier = new Supplier();
         $supplier->name = $request->name;
         $supplier->email = $request->email;
@@ -69,6 +71,7 @@ class SupplierController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize('user', 'admin', SupplierPolicy::class);
         $supplier = Supplier::findOrFail($id);
         return view('supplier.editsupplier', compact('supplier'));
     }
@@ -98,6 +101,8 @@ class SupplierController extends Controller
      */
     public function destroy($id)
     {
+        //deletion is not recommended
+        $this->authorize('user', 'admin', SupplierPolicy::class);
         $supplier = Supplier::findorFail($id);
         $supplier->delete();
         return redirect('/suppliers')->with('success', 'supplier deleted!');
